cluster::Event: Move implementation into cluster/Event.{h,cc}
